Fixed issue with sparse matrix merging, added AMAZON switch to ignore ssd write step

+++ Snakefile
@@ -58,16 +58,19 @@ rule all:
     input: expand("mapping/{sample}/{sample}/wssd_out_file", sample = SAMPLES.keys())
 
 rule merge_sparse_matrices:
-    input: expand("region_matrices/{sample}/{sample}.{part}_%d.pkl" % BAM_PARTITIONS, sample = SAMPLES.keys(), part = range(BAM_PARTITIONS))
+    input: expand("region_matrices/{{sample}}/{{sample}}.{part}_%d.pkl" % BAM_PARTITIONS, part = range(BAM_PARTITIONS))
     output: "mapping/{sample}/{sample}/wssd_out_file"
     params: sge_opts = "-l mfree=32G -l data_scratch_ssd_disk_free=10G -pe serial 1 -N merge_sample"
     log: "log/merge/{sample}.txt"
     benchmark: "benchmarks/merger/{sample}.json"
     run:
-        shell("mkdir -p /data/scratch/ssd/{wildcards.sample}")
-        shell("python3 merger.py /data/scratch/ssd/{wildcards.sample}/wssd_out_file --infiles {input}")
-        shell("rsync /data/scratch/ssd/{wildcards.sample}/wssd_out_file {output}")
-        shell("rm /data/scratch/ssd/{wildcards.sample}/wssd_out_file")
+        if AMAZON:
+            shell("python3 merger.py {output} --infiles {input}")
+        else:
+            shell("mkdir -p /data/scratch/ssd/{wildcards.sample}")
+            shell("python3 merger.py {output} --infiles {input}")
+            shell("rsync /data/scratch/ssd/{wildcards.sample}/wssd_out_file {output}")
+            shell("rm /data/scratch/ssd/{wildcards.sample}/wssd_out_file")
 
 #rule merge_matrices_live:
 #    input: lambda wildcards: SAMPLES[wildcards.sample]
